requirements.txt
APScheduler==3.5.3
basketball-reference-web-scraper==4.0.0
certifi==2018.10.15
chardet==3.0.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.1.4
django-toolbelt==0.0.1
djangorestframework==3.9.0
gunicorn==19.9.0
idna==2.7
lxml==4.2.5
psycopg2==2.7.6.1
pytz==2018.6
requests==2.20.0
six==1.12.0
static3==0.7.0
tzlocal==1.5.1
urllib3==1.24
whitenoise==4.1.2
